+++ packages/pluggableWidgets/bottom-sheet-native/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Fixed
 
+-   Fixed iOS orientation change when opening the bottom sheet modal on devices in landscape mode.
 -   Fixed flickering issue on Android when opening bottom sheet (both basic and custom render types).
 -   Improved backdrop animation with smooth fade-in/fade-out transitions.
 -   Fixed bottomsheet issue to close when the trigger attribute changes.